sysfree/requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.1
binaryornot==0.4.4
Brotli==1.1.0
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cookiecutter==2.6.0
coverage==7.8.0
cryptography==44.0.2
cssselect2==0.8.0
Django==5.2
django-appconf==1.1.0
django-auditlog==3.0.0
django-compressor==4.5.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-haystack==3.3.0
django-phonenumber-field==8.0.0
django-rosetta==0.10.2
django-simpleui==2025.1.13
django-tailwind==4.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
fonttools==4.57.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
kombu==5.5.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.6
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
paypal-checkout-serversdk==1.0.3
paypalhttp==1.0.1
paypalrestsdk==1.13.3
phonenumbers==9.0.1
pillow==11.1.0
polib==1.2.0
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg==3.2.6
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.9.0
pymemcache==4.0.0
pyOpenSSL==25.0.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
rcssmin==1.1.2
redis==6.1.0
reportlab==4.4.1
requests==2.32.3
rich==14.0.0
rjsmin==1.2.2
setuptools==79.0.1
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
Whoosh==2.7.4
XlsxWriter==3.2.3
zopfli==0.2.3.post1
